+++ ext/log-collector/src/main/java/core/log/LogCollectorApp.java
@@ -21,7 +21,9 @@ public class LogCollectorApp extends App {
     protected void initialize() {
         load(new SystemModule("sys.properties"));
         loadProperties("app.properties");
-        http().maxForwardedIPs(3);      // loose x-forwarded-for ip config, there are cdn/proxy before system, and in event collector, preventing fake client ip is less important
+        // default is 3 to loose x-forwarded-for ip config, there are cdn/proxy before system, and in event collector, preventing fake client ip is less important
+        // use env to override
+        http().maxForwardedIPs(Integer.parseInt(requiredProperty("sys.http.maxForwardedIPs")));
 
         site().security();
         site().staticContent("/robots.txt").cache(Duration.ofDays(30));