+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Change log
 
 ### 7.10.7 (04/29/2022 - )
+* log-collector: make http().maxForwardedIPs() read property "sys.http.maxForwardedIPs"
+  > so it can be configured by env
 
 ### 7.10.6 (04/19/2022 - 04/28/2022)